requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlmodel>=0.0.9
motor>=3.3.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0
pytest>=7.4.2
pytest-asyncio>=0.21.1
httpx>=0.24.1
python-dotenv>=1.0.0
redis>=5.0.0 
fastapi-cache2